Add new field type for page numbers
This sounds like an innocuous-enough commit but it's actually quite
involved. The "page" field has been replaced by two fields--"in_preface"
, and "page_number"--for the Section, TermOccurrence, and Note models.
As a result, the forms for adding notes, sections and terms all had to
be revamped to varying degrees, both cosmetically and functionally. The
templates for displaying notes/terms/sections had to be adjusted as
well.

This closes #12. Also closes #13 since now the section is determined
solely from the page (assumes that all section page numbers are filled).