Optimise database queries
Not perfect but it's a start. Had to get rid of the summary percent
thing (it wasn't useful anyway) + section couts in book_list.html (still
not sure why, but if I have 3 annotations instead of 2, it takes an
order of magnitude longer to execute the query, and the number of
sections is the least important metric so it's gone).