Fix page number field validation with postgres
Now that we're using postgres, there's some field validation going on
that I can't seem to override which conflicts with my janky preface
hack. To get around it, for now, I've just changed the python field
type to CharField, even though it's still an integer field in the actual
db. Not ideal but it seems to work.