Improve logic for adding books/authors
This fixes #98 by removing the goodreads_id field from Author, and
moving it to a separate GoodreadsAuthor model. The logic for adding
books and authors has been removed from the view and moved to the
model manager.

Comes with tests, a management command for validating existing book
data, and a management command for merging authors (as with n+1).