Change sort for goodreads sync page
The website has apparently been updated to show a max of 10 books at
a time for logged-out users, so the per_page=100 query param doesn't
do anything anymore. Can't figure out how to get it back. In the
meantime I'm changing this to sort by date read so at least it'll be
easier to find a specific book.