Fix bug with post-delete signal handler for books
This is a little tricky to explain but basically I added in a signal
handler a while back to make sure that the related bookdetails is
deleted when the book is deleted (since models.CASCADE doesn't do that
by default for OneToOneFields). But when deleting from the edit book
page, we are deleting the BookDetails and not the Book (I think it's
more logical this way since then it tells us all the objects that will
be deleted in the admin interface). So there was an error being thrown
about the BookDetails already being deleted by the time the signal
handler is run. So this seems to fix it.