Add CSRF settings for enabling https
I just did a really basic bad thing by changing the proxy setting
for bookmarker.dellsystem.me (previously it was off). So https seems
to be supported but it's not like mandated. Good enough for now I think.
This change was necessary to get form submission working again.

Relevant for #104 (but doesn't close it - I need to actually force it)